feat: 타입별 피드에서 공통되는 버튼과 프리뷰 분리
