+++ src/components/Feed/NormalFeed/NormalFeed.tsx
@@ -4,21 +4,14 @@ import Divider from '@components/common/Divider/Divider';
 import Drawer from '@components/common/Drawer/Drawer';
 import Flex from '@components/common/Flex/Flex';
 import Heading from '@components/common/Heading/Heading';
-import IconButton from '@components/common/IconButton/IconButton';
-import Text from '@components/common/Text/Text';
+import ActionButton from '@components/Feed/ActionButton/ActionButton';
 import NormalDetail from '@components/Feed/Detail/Normal/NormalDetail';
 import FeedAuthor from '@components/Feed/FeedAuthors/FeedAuthor';
+import { CommentPreview } from '@components/Feed/Preview/Preview';
 import { getFormattedDate } from '@utils/getFormattedDate';
 import { FEED_DETAIL_MAX_HEIGHT } from '@styles/layout';
 import type { FeedData } from '@type/feed';
 import * as S from './NormalFeed.styled';
-
-interface ActionButtonProps {
-	icon: 'Comment' | 'Attachment';
-	count: number;
-	onClick: () => void;
-	ariaLabel: string;
-}
 
 interface FeedProps {
 	feedData: FeedData;
@@ -26,49 +19,6 @@ interface FeedProps {
 	openDetail: () => void;
 	closeDetail: () => void;
 }
-
-const ActionButton = ({
-	icon,
-	count,
-	onClick,
-	ariaLabel,
-}: ActionButtonProps) => {
-	return (
-		<Flex align='center'>
-			<IconButton
-				icon={icon}
-				size='md'
-				color='secondary'
-				ariaLabel={ariaLabel}
-				onClick={onClick}
-			/>
-			<Text color='secondary' size='md' weight='medium'>
-				{count === 0 ? '0' : String(count)}
-			</Text>
-		</Flex>
-	);
-};
-
-const CommentPreview = ({ comments }: { comments: FeedData['comments'] }) => {
-	if (comments.length === 0) return null;
-
-	const commentsToShow =
-		comments.length === 1 ? comments.slice(-1) : comments.slice(-2);
-	return (
-		<Flex direction='column' gap='8' marginBottom='10'>
-			{commentsToShow.map((comment) => (
-				<Flex key={comment.id} gap='6'>
-					<Text size='md' weight='semiBold'>
-						{comment.author.username}
-					</Text>
-					<Text size='md' weight='regular'>
-						{comment.content}
-					</Text>
-				</Flex>
-			))}
-		</Flex>
-	);
-};
 
 const NormalFeed = ({
 	feedData,
